// packages/react-line-flex/src/types.ts
export type FlexSpacing = 'none' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
export type FlexMargin = FlexSpacing;
export type FlexGravity = 'top' | 'bottom' | 'center';
export type FlexAlign = 'start' | 'end' | 'center';
export type FlexDecoration = 'none' | 'underline' | 'line-through';
export type FlexWeight = 'regular' | 'bold';
export type FlexStyle = 'normal' | 'italic';
export type FlexPosition = 'relative' | 'absolute';
// ...
export type FlexButtonStyle = 'link' | 'primary' | 'secondary';
export type FlexButtonHeight = 'sm' | 'md';
export type FlexBorderWidth = 'none' | 'light' | 'normal' | 'medium' | 'semi-bold' | 'bold';

export type FlexAction = FlexURIAction | FlexMessageAction | FlexPostbackAction;

export interface FlexURIAction {
  type: 'uri';
  label?: string;
  uri: string;
  altUri?: {
    desktop?: string;
  };
}

export interface FlexMessageAction {
  type: 'message';
  label?: string;
  text: string;
}

export interface FlexPostbackAction {
  type: 'postback';
  label?: string;
  data: string;
  displayText?: string;
}
// ...
export interface FlexButton {
  type: 'button';
  action: FlexAction;
  flex?: number;
  margin?: FlexMargin | string;
  position?: FlexPosition;
  offsetTop?: string;
  offsetBottom?: string;
  offsetStart?: string;
  offsetEnd?: string;
  height?: FlexButtonHeight;
  style?: FlexButtonStyle;
  color?: string;
  gravity?: FlexGravity;
  adjustMode?: 'shrink-to-fit';
}